CODE RUB: Remove Dublicate Time Folders
