--- Tarteeb.Api/Models/Foundations/TimeSlots/Exceptions/FailedTimeServiceException.cs
@@ -1,17 +1,0 @@
-﻿//=================================
-// Copyright (c) Coalition of Good-Hearted Engineers
-// Free to use to bring order in your workplace
-//=================================
-
-using System;
-using Xeptions;
-
-namespace Tarteeb.Api.Models.Foundations.TimeSlots.Exceptions
-{
-    public class FailedTimeServiceException : Xeption
-    {
-        public FailedTimeServiceException(Exception innerException)
-            : base(message: "Failed time service error occurred, please contact support", innerException)
-        { }
-    }
-}
