--- Tarteeb.Api/Models/Foundations/TimeSlots/Exceptions/TimeDependencyException.cs
@@ -1,16 +1,0 @@
-﻿//=================================
-// Copyright (c) Coalition of Good-Hearted Engineers
-// Free to use to bring order in your workplace
-//=================================
-
-using Xeptions;
-
-namespace Tarteeb.Api.Models.Foundations.TimeSlots.Exceptions
-{
-    public class TimeDependencyException : Xeption
-    {
-        public TimeDependencyException(Xeption innerException)
-           : base(message: "Time dependency error occurred, contact support.", innerException)
-        { }
-    }
-}
