--- Tarteeb.Api/Models/Foundations/TimeSlots/Exceptions/TimeValidationException.cs
@@ -1,16 +1,0 @@
-﻿//=================================
-// Copyright (c) Coalition of Good-Hearted Engineers
-// Free to use to bring order in your workplace
-//=================================
-
-using Xeptions;
-
-namespace Tarteeb.Api.Models.Foundations.TimeSlots.Exceptions
-{
-    public class TimeValidationException : Xeption
-    {
-        public TimeValidationException(Xeption innerException)
-            : base(message: "Time validation error occurred, fix the errors and try again.", innerException)
-        { }
-    }
-}
