+++ Tarteeb.Api/Models/Foundations/Times/Exceptions/FailedTimeServiceException.cs
@@ -11,8 +11,7 @@ namespace Tarteeb.Api.Models.Foundations.Times.Exceptions
     public class FailedTimeServiceException : Xeption
     {
         public FailedTimeServiceException(Exception innerException)
-            : base(message: "Failed time service error occurred, please contact support.",
-                innerException)
+            : base(message: "Failed time service error occurred, please contact support", innerException)
         { }
     }
 }
