+++ Tarteeb.Api/Models/Foundations/Times/Exceptions/NullTimeException.cs
@@ -9,7 +9,7 @@ namespace Tarteeb.Api.Models.Foundations.Times.Exceptions
 {
     public class NullTimeException : Xeption
     {
-        public NullTimeException()
-        : base(message: "Time is null") { }
+        public NullTimeException() : base("Time is null.")
+        { }
     }
 }
