+++ Tarteeb.Api/Models/Foundations/Times/Exceptions/TimeServiceException.cs
@@ -3,14 +3,13 @@
 // Free to use to bring order in your workplace
 //=================================
 
-using System;
 using Xeptions;
 
 namespace Tarteeb.Api.Models.Foundations.Times.Exceptions
 {
     public class TimeServiceException : Xeption
     {
-        public TimeServiceException(Exception innerException)
+        public TimeServiceException(Xeption innerException)
             : base(message: "Time service error occurred, contact support.", innerException)
         { }
     }
