+++ Tarteeb.Api/Models/Foundations/Times/Exceptions/TimeValidationException.cs
@@ -10,7 +10,7 @@ namespace Tarteeb.Api.Models.Foundations.Times.Exceptions
     public class TimeValidationException : Xeption
     {
         public TimeValidationException(Xeption innerException)
-            : base(message: "Time validation error occurred, fix the errors and try again", innerException)
+            : base(message: "Time validation error occurred, fix the errors and try again.", innerException)
         { }
     }
 }
